new input handler routine with network buffer support
this new routine completely replaces the game's original input handler routine. It creates and manages a stack of the inputs from previous frames for players 1 and 2. The stack size is adjustable. Inputs are written to the stack with a timestamp and are executed based on the timestamp of when they were created, relative to the adjustable input buffer provided for each player, agnostic of the input's position in the stack. The netcode script is in charge of writing player 2's inputs to their stack.